Fix zion-link-to-src
zion-link-to-src was running exit in a subshell, so it didn't actually exit.
This caused it to delete lib and runtime from the current directory.  Switch to
a compound statement instead of a subshell.

Also pass in ZION_ROOT pointing into DESTDIR so that it doesn't attempt to
update a hardcoded zion path.